Add Types class
