+++ Game-of-Life/Cells/BaseCell.cs
@@ -1,4 +1,5 @@
 ﻿using System.Windows.Controls;
+using Game_of_Life.Options;
 
 namespace Game_of_Life.Cells
 {
@@ -26,7 +27,7 @@ namespace Game_of_Life.Cells
         //private int x, y;
         public Image Model;                   // Объект картинки. Нужен для визуального отображения пустой клетки
 
-        public EmptyCells.CellType CellType;  // Объект перечисления, который хранит название типа хлетки, которая находится в этой (пустой) клетке
+        public CellType cellType;  // Объект перечисления, который хранит название типа хлетки, которая находится в этой (пустой) клетке
 
         //-------------------------------------------------<Конструкторы>---------------------------------------------------------------------------------\\
