+++ Game-of-Life/Options/Settings.cs
@@ -20,8 +20,8 @@
          *
          */
 
-        static public int ScrollMoveSpeed;
+        static public int ScrollMoveSpeed = 6;
 
-        static public int ScrollDistSpeed;
+        static public int ScrollDistSpeed = 6;
     }
 }
